feat(iast): add secure marks core and sanitize cmdi
