+++ tests/appsec/iast/fixtures/taint_sinks/sql_injection_psycopg2.py
@@ -1,25 +1,8 @@
-import os
-
-import psycopg2
 from psycopg2.errors import DuplicateTable
 
 from ddtrace.appsec._iast._taint_tracking._taint_objects import get_tainted_ranges
 from ddtrace.appsec._iast._taint_tracking._taint_objects import is_pyobject_tainted
-
-
-POSTGRES_HOST = os.getenv("TEST_POSTGRES_HOST", "127.0.0.1")
-
-
-def get_connection():
-    connection = psycopg2.connect(
-        user="postgres", password="postgres", host=POSTGRES_HOST, port="5432", database="postgres"
-    )
-    return connection
-
-
-def close_connection(connection):
-    if connection:
-        connection.close()
+from tests.appsec.iast.db_utils import get_connection
 
 
 def sqli_simple(table):
